Add: ComboController, BookingController. Update: Index of table

+++ BE-v2/PartyKid/Controllers/ServiceCategoriesController.cs
@@ -42,7 +42,7 @@ public class ServiceCategoriesController : BaseApi
 
     [Authorize(nameof(RoleCollection.Admin))]
     [HttpPost]
-    public async Task<IResponse> Create([FromBody] CreateServiceBindingModel request)
+    public async Task<IResponse> Create([FromBody] CreateServiceCategoryBindingModel request)
     {
         string result = await _serviceCategoryService.Create(_mapper.Map<ServiceCategory>(request));
         return Success(message: result);
@@ -51,7 +51,7 @@ public class ServiceCategoriesController : BaseApi
     [Authorize(nameof(RoleCollection.Admin))]
     [HttpPut]
     [Route("{Id}")]
-    public async Task<IResponse> Update([FromRoute(Name = "Id")] int id, [FromBody] UpdateServiceBindingModel request)
+    public async Task<IResponse> Update([FromRoute(Name = "Id")] int id, [FromBody] UpdateServiceCategoryBindingModel request)
     {
         ServiceCategory serviceCategory = await _serviceCategoryService.Find(id);
         if (serviceCategory is null)
@@ -69,6 +69,11 @@ public class ServiceCategoriesController : BaseApi
     [Route("{Id}")]
     public async Task<IResponse> Delete([FromRoute(Name = "Id")] int id)
     {
+        ServiceCategory serviceCategory = await _serviceCategoryService.Find(id);
+        if (serviceCategory is null)
+        {
+            throw new DomainException(Constants.Transactions.Messages.NotFound);
+        }
         return Success(message: await _serviceCategoryService.Delete(id));
     }
     #endregion
